+++ package.json
@@ -5,7 +5,7 @@
   "main": "index.js",
   "scripts": {
     "start": "node index.js",
-    "test": "echo \"Error: no test specified\" && exit 1",
+    "test": "node --experimental-vm-modules node_modules/jest/bin/jest.js",
     "lint": "prettier --check .",
     "lint:fix": "prettier --write ."
   },
@@ -18,6 +18,8 @@
     "@octokit/webhooks": "^10.0.8"
   },
   "devDependencies": {
+    "jest": "^28.1.3",
     "prettier": "2.7.1"
-  }
+  },
+  "transform": {}
 }
